adding dependency to the incubator project , refactoring the configuration to create an HTTP3 version of it
