+++ gradle.properties
@@ -3,5 +3,6 @@ version=1.1.14-SNAPSHOT
 reactorNettyQuicVersion=0.1.14-SNAPSHOT
 reactorCoreVersion=3.5.13-SNAPSHOT
 reactorAddonsVersion=3.5.2-SNAPSHOT
+nettyHttp3Version=0.0.23.Final-SNAPSHOT
 compatibleVersion=1.1.13
 bomVersion=2022.0.13
